Update views.py
It seems there's a bug on the database. The website seems to be storing more than 1 entrance to each day, so when you just add the number of transactions of a given month, the total transactions are higher than the real number. So the chart for shielded transactions is showing wrong numbers. I tried this version and it worked fine, so I'm updating it. Give it a try and check to see if monthly numbers are equal to mine on https://moneroj.net/transmonth/